refactor(frontend): modularize profile feature with dedicated widgets and screens
