+++ frontend/lib/screens/profile/progress_tab.dart
@@ -1,27 +1,28 @@
 import 'package:flutter/material.dart';
 import 'package:syntrak/core/theme.dart';
 import 'package:syntrak/models/activity.dart';
-import 'package:intl/intl.dart';
-import 'package:table_calendar/table_calendar.dart';
+import 'package:syntrak/screens/profile/widgets/progress/progress_activity_calendar.dart';
+import 'package:syntrak/screens/profile/widgets/progress/progress_insight_cards.dart';
+import 'package:syntrak/screens/profile/widgets/progress/progress_streaks_banner.dart';
+import 'package:syntrak/screens/profile/widgets/progress/progress_weekly_overview.dart';
 
 class ProgressTab extends StatefulWidget {
-  final List<Activity> activities;
-
   const ProgressTab({
     super.key,
     required this.activities,
   });
+
+  final List<Activity> activities;
 
   @override
   State<ProgressTab> createState() => _ProgressTabState();
 }
 
 class _ProgressTabState extends State<ProgressTab> {
-  // Mock data - will be replaced with backend data
   final Map<String, dynamic> _weeklyStats = {
-    'distance': 0.0, // km
-    'time': 0, // minutes
-    'elevGain': 0.0, // meters
+    'distance': 0.0,
+    'time': 0,
+    'elevGain': 0.0,
   };
 
   final List<Map<String, dynamic>> _bestEfforts = [];
@@ -34,11 +35,11 @@ class _ProgressTabState extends State<ProgressTab> {
     'previous': 22,
   };
   final Map<String, dynamic> _trainingLog = {
-    'distance': 10.9, // km
+    'distance': 10.9,
     'dateRange': 'Jan 5 - Jan 11, 2026',
   };
 
-  final Set<DateTime> _activityDays = {}; // Days with activities
+  final Set<DateTime> _activityDays = {};
 
   @override
   void initState() {
@@ -47,7 +48,6 @@ class _ProgressTabState extends State<ProgressTab> {
   }
 
   void _calculateStats() {
-    // Calculate weekly stats
     final now = DateTime.now();
     final weekStart = now.subtract(Duration(days: now.weekday - 1));
 
@@ -55,10 +55,13 @@ class _ProgressTabState extends State<ProgressTab> {
     int weeklyTime = 0;
     double weeklyElevGain = 0;
 
-    for (var activity in widget.activities) {
+    _bestEfforts.clear();
+    _activityDays.clear();
+
+    for (final activity in widget.activities) {
       if (activity.startTime.isAfter(weekStart)) {
-        weeklyDistance += activity.distance / 1000; // Convert to km
-        weeklyTime += activity.duration ~/ 60; // Convert to minutes
+        weeklyDistance += activity.distance / 1000;
+        weeklyTime += activity.duration ~/ 60;
         weeklyElevGain += activity.elevationGain;
         _activityDays.add(DateTime(
           activity.startTime.year,
@@ -68,9 +71,7 @@ class _ProgressTabState extends State<ProgressTab> {
       }
     }
 
-    // Calculate best efforts (mock for now)
     if (widget.activities.isNotEmpty) {
-      // Sort by distance, time, etc. to find best efforts
       final sortedByDistance = List<Activity>.from(widget.activities)
         ..sort((a, b) => b.distance.compareTo(a.distance));
 
@@ -95,7 +96,6 @@ class _ProgressTabState extends State<ProgressTab> {
   Widget build(BuildContext context) {
     return RefreshIndicator(
       onRefresh: () async {
-        // TODO: Refresh data from backend
         await Future.delayed(const Duration(seconds: 1));
       },
       color: SyntrakColors.primary,
@@ -103,814 +103,23 @@ class _ProgressTabState extends State<ProgressTab> {
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            // Streaks subsection
-            _buildStreaksSection(),
-
-            // Weekly activity diagram
-            _buildWeeklyActivityDiagram(),
-
+            const ProgressStreaksBanner(),
+            ProgressWeeklyOverview(
+              weeklyStats: _weeklyStats,
+              activities: widget.activities,
+            ),
             const SizedBox(height: SyntrakSpacing.lg),
-
-            // Best Efforts card
-            _buildBestEffortsCard(),
-
-            const SizedBox(height: SyntrakSpacing.md),
-
-            // Goals card
-            _buildGoalsCard(),
-
-            const SizedBox(height: SyntrakSpacing.md),
-
-            // Relative Effort card
-            _buildRelativeEffortCard(),
-
-            const SizedBox(height: SyntrakSpacing.md),
-
-            // Training Log card
-            _buildTrainingLogCard(),
-
-            const SizedBox(height: SyntrakSpacing.lg),
-
-            // Start free trial button
-            _buildFreeTrialButton(),
-
-            const SizedBox(height: SyntrakSpacing.xl),
-
-            // Activity calendar
-            _buildActivityCalendar(),
-
+            ProgressInsightCards(
+              bestEfforts: _bestEfforts,
+              goals: _goals,
+              relativeEffort: _relativeEffort,
+              trainingLog: _trainingLog,
+            ),
+            ProgressActivityCalendar(activityDays: _activityDays),
             const SizedBox(height: SyntrakSpacing.xl),
           ],
         ),
       ),
     );
   }
-
-  Widget _buildStreaksSection() {
-    return Container(
-      margin: const EdgeInsets.all(SyntrakSpacing.md),
-      padding: const EdgeInsets.all(SyntrakSpacing.md),
-      decoration: BoxDecoration(
-        color: SyntrakColors.surfaceVariant,
-        borderRadius: BorderRadius.circular(SyntrakRadius.lg),
-      ),
-      child: Row(
-        children: [
-          Icon(
-            Icons.local_fire_department,
-            color: SyntrakColors.accent,
-            size: 24,
-          ),
-          const SizedBox(width: SyntrakSpacing.md),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  'Scroll down for streaks',
-                  style: SyntrakTypography.labelLarge.copyWith(
-                    color: SyntrakColors.textPrimary,
-                    fontWeight: FontWeight.w600,
-                  ),
-                ),
-                const SizedBox(height: SyntrakSpacing.xs),
-                Text(
-                  'Build habits with streaks - log one activity a week to keep it alive',
-                  style: SyntrakTypography.bodySmall.copyWith(
-                    color: SyntrakColors.textSecondary,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildWeeklyActivityDiagram() {
-    return Container(
-      margin: const EdgeInsets.symmetric(horizontal: SyntrakSpacing.md),
-      padding: const EdgeInsets.all(SyntrakSpacing.md),
-      decoration: BoxDecoration(
-        color: SyntrakColors.surface,
-        borderRadius: BorderRadius.circular(SyntrakRadius.lg),
-        border: Border.all(color: SyntrakColors.divider),
-        boxShadow: SyntrakElevation.sm,
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            'This week',
-            style: SyntrakTypography.headlineMedium.copyWith(
-              color: SyntrakColors.textPrimary,
-            ),
-          ),
-          const SizedBox(height: SyntrakSpacing.md),
-          // Weekly stats
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceAround,
-            children: [
-              Expanded(
-                child: _buildStatItem(
-                  'Distance',
-                  '${_weeklyStats['distance'].toStringAsFixed(1)} km',
-                ),
-              ),
-              Expanded(
-                child: _buildStatItem(
-                  'Time',
-                  '${_weeklyStats['time']}m',
-                ),
-              ),
-              Expanded(
-                child: _buildStatItem(
-                  'Elev Gain',
-                  '${_weeklyStats['elevGain'].toStringAsFixed(0)} m',
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: SyntrakSpacing.lg),
-          // Past 12 weeks graph
-          Text(
-            'Past 12 weeks',
-            style: SyntrakTypography.bodyLarge.copyWith(
-              color: SyntrakColors.textPrimary,
-            ),
-          ),
-          const SizedBox(height: SyntrakSpacing.sm),
-          SizedBox(
-            height: 140,
-            child: _build12WeeksGraph(),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildStatItem(String label, String value) {
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        Text(
-          value,
-          style: SyntrakTypography.headlineSmall.copyWith(
-            color: SyntrakColors.textPrimary,
-          ),
-          textAlign: TextAlign.center,
-          maxLines: 1,
-          overflow: TextOverflow.ellipsis,
-        ),
-        const SizedBox(height: SyntrakSpacing.xs),
-        Text(
-          label,
-          style: SyntrakTypography.labelSmall.copyWith(
-            color: SyntrakColors.textTertiary,
-          ),
-          textAlign: TextAlign.center,
-          maxLines: 1,
-          overflow: TextOverflow.ellipsis,
-        ),
-      ],
-    );
-  }
-
-  Widget _build12WeeksGraph() {
-    // Calculate weekly distances from activities
-    final now = DateTime.now();
-    final weeks = List.generate(12, (index) {
-      final weekStart =
-          now.subtract(Duration(days: (11 - index) * 7 + now.weekday - 1));
-      final weekEnd = weekStart.add(const Duration(days: 6));
-
-      double weekDistance = 0.0;
-      for (var activity in widget.activities) {
-        if (activity.startTime
-                .isAfter(weekStart.subtract(const Duration(days: 1))) &&
-            activity.startTime.isBefore(weekEnd.add(const Duration(days: 1)))) {
-          weekDistance += activity.distance / 1000; // Convert to km
-        }
-      }
-
-      return {
-        'date': weekStart,
-        'distance': weekDistance,
-      };
-    });
-
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        SizedBox(
-          height: 100,
-          child: CustomPaint(
-            painter: _GraphPainter(weeks),
-            child: Container(),
-          ),
-        ),
-        const SizedBox(height: SyntrakSpacing.xs),
-        // Month labels
-        Padding(
-          padding: const EdgeInsets.symmetric(horizontal: SyntrakSpacing.xs),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              if (weeks.isNotEmpty)
-                Text(
-                  DateFormat('MMM').format(weeks[0]['date'] as DateTime),
-                  style: SyntrakTypography.labelSmall.copyWith(
-                    color: SyntrakColors.textTertiary,
-                  ),
-                ),
-              if (weeks.length > 6)
-                Text(
-                  DateFormat('MMM').format(weeks[6]['date'] as DateTime),
-                  style: SyntrakTypography.labelSmall.copyWith(
-                    color: SyntrakColors.textTertiary,
-                  ),
-                ),
-              if (weeks.length > 11)
-                Text(
-                  DateFormat('MMM').format(weeks[11]['date'] as DateTime),
-                  style: SyntrakTypography.labelSmall.copyWith(
-                    color: SyntrakColors.textTertiary,
-                  ),
-                ),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildBestEffortsCard() {
-    return _buildCard(
-      title: 'Best Efforts',
-      child: _bestEfforts.isEmpty
-          ? Padding(
-              padding: const EdgeInsets.symmetric(
-                horizontal: SyntrakSpacing.md,
-                vertical: SyntrakSpacing.lg,
-              ),
-              child: Text(
-                'No best efforts yet. Complete activities to see your records!',
-                style: SyntrakTypography.bodyMedium.copyWith(
-                  color: SyntrakColors.textTertiary,
-                ),
-                textAlign: TextAlign.center,
-              ),
-            )
-          : Column(
-              mainAxisSize: MainAxisSize.min,
-              children: _bestEfforts.map((effort) {
-                return _buildBestEffortItem(
-                  isPR: effort['isPR'] ?? false,
-                  type: effort['type'] ?? '',
-                  time: effort['time'] ?? '',
-                  date: effort['date'] as DateTime? ?? DateTime.now(),
-                );
-              }).toList(),
-            ),
-    );
-  }
-
-  Widget _buildBestEffortItem({
-    required bool isPR,
-    required String type,
-    required String time,
-    required DateTime date,
-  }) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(
-        horizontal: SyntrakSpacing.md,
-        vertical: SyntrakSpacing.sm,
-      ),
-      child: Row(
-        children: [
-          Container(
-            width: 40,
-            height: 40,
-            decoration: BoxDecoration(
-              color: isPR
-                  ? SyntrakColors.accent.withOpacity(0.2)
-                  : SyntrakColors.textTertiary.withOpacity(0.2),
-              shape: BoxShape.circle,
-            ),
-            child: Icon(
-              isPR ? Icons.emoji_events : Icons.military_tech,
-              color: isPR ? SyntrakColors.accent : SyntrakColors.textTertiary,
-              size: 20,
-            ),
-          ),
-          const SizedBox(width: SyntrakSpacing.md),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Row(
-                  children: [
-                    Text(
-                      isPR ? 'PR' : '2nd-fastest',
-                      style: SyntrakTypography.labelMedium.copyWith(
-                        color: SyntrakColors.textPrimary,
-                        fontWeight: FontWeight.w600,
-                      ),
-                    ),
-                    const SizedBox(width: SyntrakSpacing.xs),
-                    Flexible(
-                      child: Text(
-                        type,
-                        style: SyntrakTypography.bodyMedium.copyWith(
-                          color: SyntrakColors.textSecondary,
-                        ),
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                    ),
-                  ],
-                ),
-                const SizedBox(height: SyntrakSpacing.xs / 2),
-                Text(
-                  DateFormat('d MMM yyyy').format(date),
-                  style: SyntrakTypography.bodySmall.copyWith(
-                    color: SyntrakColors.textTertiary,
-                  ),
-                ),
-              ],
-            ),
-          ),
-          Text(
-            time,
-            style: SyntrakTypography.bodyLarge.copyWith(
-              color: SyntrakColors.textPrimary,
-              fontWeight: FontWeight.w600,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildGoalsCard() {
-    return _buildCard(
-      title: 'Goals',
-      child: Padding(
-        padding: const EdgeInsets.symmetric(
-          horizontal: SyntrakSpacing.md,
-          vertical: SyntrakSpacing.md,
-        ),
-        child: Row(
-          children: [
-            Stack(
-              alignment: Alignment.center,
-              children: [
-                SizedBox(
-                  width: 60,
-                  height: 60,
-                  child: CircularProgressIndicator(
-                    value: _goals['weeklyRuns']['current'] /
-                        _goals['weeklyRuns']['target'],
-                    strokeWidth: 4,
-                    backgroundColor: SyntrakColors.surfaceVariant,
-                    valueColor: AlwaysStoppedAnimation<Color>(
-                      SyntrakColors.success,
-                    ),
-                  ),
-                ),
-                Icon(
-                  Icons.downhill_skiing,
-                  color: SyntrakColors.primary,
-                  size: 24,
-                ),
-              ],
-            ),
-            const SizedBox(width: SyntrakSpacing.md),
-            Expanded(
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    _goals['description'],
-                    style: SyntrakTypography.bodyMedium.copyWith(
-                      color: SyntrakColors.textPrimary,
-                    ),
-                    maxLines: 2,
-                    overflow: TextOverflow.ellipsis,
-                  ),
-                  const SizedBox(height: SyntrakSpacing.xs),
-                  Text(
-                    '${_goals['weeklyRuns']['current']}/${_goals['weeklyRuns']['target']} activities',
-                    style: SyntrakTypography.bodySmall.copyWith(
-                      color: SyntrakColors.textSecondary,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-            Icon(
-              Icons.chevron_right,
-              color: SyntrakColors.textTertiary,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildRelativeEffortCard() {
-    return _buildCard(
-      title: 'Relative Effort',
-      child: Padding(
-        padding: const EdgeInsets.symmetric(vertical: SyntrakSpacing.sm),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            _buildRelativeEffortItem(
-              _relativeEffort['current'],
-              'Jan 6 - Jan 12, 2026',
-              SyntrakColors.error,
-            ),
-            Divider(
-              height: 1,
-              color: SyntrakColors.divider,
-            ),
-            _buildRelativeEffortItem(
-              _relativeEffort['previous'],
-              'Dec 30 - Jan 5, 2026',
-              SyntrakColors.snowboard,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildRelativeEffortItem(int value, String dateRange, Color color) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(
-        horizontal: SyntrakSpacing.md,
-        vertical: SyntrakSpacing.sm,
-      ),
-      child: Row(
-        children: [
-          Container(
-            width: 40,
-            height: 40,
-            decoration: BoxDecoration(
-              color: color.withOpacity(0.2),
-              borderRadius: BorderRadius.circular(SyntrakRadius.md),
-            ),
-            child: Center(
-              child: Text(
-                value.toString(),
-                style: SyntrakTypography.headlineSmall.copyWith(
-                  color: color,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-            ),
-          ),
-          const SizedBox(width: SyntrakSpacing.md),
-          Expanded(
-            child: Text(
-              dateRange,
-              style: SyntrakTypography.bodyMedium.copyWith(
-                color: SyntrakColors.textPrimary,
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildTrainingLogCard() {
-    return _buildCard(
-      title: 'Training Log',
-      child: Padding(
-        padding: const EdgeInsets.symmetric(
-          horizontal: SyntrakSpacing.md,
-          vertical: SyntrakSpacing.md,
-        ),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                Expanded(
-                  child: Text(
-                    _trainingLog['dateRange'],
-                    style: SyntrakTypography.bodyMedium.copyWith(
-                      color: SyntrakColors.textPrimary,
-                    ),
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                  ),
-                ),
-                Text(
-                  '${_trainingLog['distance']} km',
-                  style: SyntrakTypography.headlineSmall.copyWith(
-                    color: SyntrakColors.textPrimary,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: SyntrakSpacing.md),
-            // Calendar week view
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceAround,
-              children: ['M', 'T', 'W', 'T', 'F', 'S', 'S'].map((day) {
-                return SizedBox(
-                  width: 30,
-                  child: Center(
-                    child: Text(
-                      day,
-                      style: SyntrakTypography.labelSmall.copyWith(
-                        color: SyntrakColors.textTertiary,
-                      ),
-                    ),
-                  ),
-                );
-              }).toList(),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildFreeTrialButton() {
-    return Container(
-      margin: const EdgeInsets.symmetric(horizontal: SyntrakSpacing.md),
-      width: double.infinity,
-      child: ElevatedButton(
-        onPressed: () {
-          // TODO: Implement free trial
-        },
-        style: ElevatedButton.styleFrom(
-          backgroundColor: SyntrakColors.accent,
-          foregroundColor: SyntrakColors.textOnPrimary,
-          padding: const EdgeInsets.symmetric(vertical: SyntrakSpacing.md),
-        ),
-        child: Text(
-          'Start a free trial',
-          style: SyntrakTypography.labelLarge,
-        ),
-      ),
-    );
-  }
-
-  Widget _buildActivityCalendar() {
-    final now = DateTime.now();
-    final focusedDay = DateTime(now.year, now.month, now.day);
-
-    return Container(
-      margin: const EdgeInsets.symmetric(horizontal: SyntrakSpacing.md),
-      padding: const EdgeInsets.all(SyntrakSpacing.lg),
-      decoration: BoxDecoration(
-        color: SyntrakColors.surface,
-        borderRadius: BorderRadius.circular(SyntrakRadius.lg),
-        border: Border.all(color: SyntrakColors.divider),
-        boxShadow: SyntrakElevation.sm,
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            DateFormat('MMMM yyyy').format(now),
-            style: SyntrakTypography.headlineSmall.copyWith(
-              color: SyntrakColors.textPrimary,
-            ),
-          ),
-          const SizedBox(height: SyntrakSpacing.md),
-          TableCalendar(
-            firstDay: DateTime.utc(2020, 1, 1),
-            lastDay: DateTime.utc(2030, 12, 31),
-            focusedDay: focusedDay,
-            calendarFormat: CalendarFormat.month,
-            startingDayOfWeek: StartingDayOfWeek.monday,
-            headerVisible: false,
-            daysOfWeekVisible: true,
-            eventLoader: (date) {
-              final normalizedDate = DateTime(date.year, date.month, date.day);
-              return _activityDays.contains(normalizedDate) ? [1] : [];
-            },
-            calendarStyle: CalendarStyle(
-              outsideDaysVisible: false,
-              defaultTextStyle: SyntrakTypography.labelSmall.copyWith(
-                color: SyntrakColors.textSecondary,
-              ),
-              weekendTextStyle: SyntrakTypography.labelSmall.copyWith(
-                color: SyntrakColors.textSecondary,
-              ),
-              selectedTextStyle: SyntrakTypography.labelSmall.copyWith(
-                color: SyntrakColors.primary,
-                fontWeight: FontWeight.w600,
-              ),
-              todayTextStyle: SyntrakTypography.labelSmall.copyWith(
-                color: SyntrakColors.primary,
-                fontWeight: FontWeight.w600,
-              ),
-              todayDecoration: BoxDecoration(
-                shape: BoxShape.circle,
-                border: Border.all(
-                  color: SyntrakColors.primary,
-                  width: 2,
-                ),
-              ),
-              selectedDecoration: BoxDecoration(
-                color: SyntrakColors.primary.withOpacity(0.2),
-                shape: BoxShape.circle,
-              ),
-              markerDecoration: BoxDecoration(
-                color: SyntrakColors.primary,
-                shape: BoxShape.circle,
-              ),
-              markerSize: 4,
-              markerMargin: const EdgeInsets.only(bottom: 4),
-              cellMargin: const EdgeInsets.all(4),
-            ),
-            daysOfWeekStyle: DaysOfWeekStyle(
-              weekdayStyle: SyntrakTypography.labelSmall.copyWith(
-                color: SyntrakColors.textTertiary,
-              ),
-              weekendStyle: SyntrakTypography.labelSmall.copyWith(
-                color: SyntrakColors.textTertiary,
-              ),
-            ),
-            selectedDayPredicate: (date) {
-              return false; // Don't show selection
-            },
-            onDaySelected: (selectedDay, focusedDay) {
-              // Optional: Handle day selection
-            },
-            // Custom builder for days with activities
-            calendarBuilders: CalendarBuilders(
-              defaultBuilder: (context, date, _) {
-                final normalizedDate =
-                    DateTime(date.year, date.month, date.day);
-                final hasActivity = _activityDays.contains(normalizedDate);
-                final isToday = date.year == now.year &&
-                    date.month == now.month &&
-                    date.day == now.day;
-
-                return Container(
-                  margin: const EdgeInsets.all(4),
-                  decoration: BoxDecoration(
-                    color: hasActivity
-                        ? SyntrakColors.primary.withOpacity(0.2)
-                        : Colors.transparent,
-                    shape: BoxShape.circle,
-                    border: isToday
-                        ? Border.all(color: SyntrakColors.primary, width: 2)
-                        : null,
-                  ),
-                  child: Center(
-                    child: Text(
-                      date.day.toString(),
-                      style: SyntrakTypography.labelSmall.copyWith(
-                        color: hasActivity
-                            ? SyntrakColors.primary
-                            : SyntrakColors.textSecondary,
-                        fontWeight:
-                            isToday ? FontWeight.w600 : FontWeight.normal,
-                      ),
-                    ),
-                  ),
-                );
-              },
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildCard({required String title, required Widget child}) {
-    return Container(
-      margin: const EdgeInsets.symmetric(horizontal: SyntrakSpacing.md),
-      decoration: BoxDecoration(
-        color: SyntrakColors.surface,
-        borderRadius: BorderRadius.circular(SyntrakRadius.lg),
-        border: Border.all(color: SyntrakColors.divider),
-        boxShadow: SyntrakElevation.sm,
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Padding(
-            padding: const EdgeInsets.fromLTRB(
-              SyntrakSpacing.md,
-              SyntrakSpacing.md,
-              SyntrakSpacing.md,
-              SyntrakSpacing.sm,
-            ),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                Text(
-                  title,
-                  style: SyntrakTypography.headlineSmall.copyWith(
-                    color: SyntrakColors.textPrimary,
-                  ),
-                ),
-                Icon(
-                  Icons.chevron_right,
-                  color: SyntrakColors.textTertiary,
-                  size: 20,
-                ),
-              ],
-            ),
-          ),
-          child,
-          const SizedBox(height: SyntrakSpacing.sm),
-        ],
-      ),
-    );
-  }
-}
-
-// Custom painter for the 12 weeks graph
-class _GraphPainter extends CustomPainter {
-  final List<Map<String, dynamic>> weeks;
-
-  _GraphPainter(this.weeks);
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final paint = Paint()
-      ..color = SyntrakColors.accent
-      ..strokeWidth = 2
-      ..style = PaintingStyle.stroke;
-
-    final pointPaint = Paint()
-      ..color = SyntrakColors.accent
-      ..style = PaintingStyle.fill;
-
-    final distances = weeks.map((w) => w['distance'] as double).toList();
-    final maxDistance =
-        distances.isEmpty ? 1.0 : distances.reduce((a, b) => a > b ? a : b);
-
-    // Prevent division by zero
-    final stepX = weeks.length <= 1 ? 0.0 : size.width / (weeks.length - 1);
-    final points = <Offset>[];
-
-    for (int i = 0; i < weeks.length; i++) {
-      final distance = weeks[i]['distance'] as double;
-      final normalizedDistance =
-          maxDistance > 0 ? (distance / maxDistance) : 0.0;
-      final y = size.height - (normalizedDistance * size.height);
-      final x = weeks.length <= 1 ? size.width / 2 : i * stepX;
-
-      // Ensure valid coordinates
-      if (x.isFinite && y.isFinite) {
-        points.add(Offset(x, y));
-      }
-    }
-
-    // Draw line (only if we have valid points)
-    for (int i = 0; i < points.length - 1; i++) {
-      if (points[i].dx.isFinite &&
-          points[i].dy.isFinite &&
-          points[i + 1].dx.isFinite &&
-          points[i + 1].dy.isFinite) {
-        canvas.drawLine(points[i], points[i + 1], paint);
-      }
-    }
-
-    // Draw points (only valid ones)
-    for (final point in points) {
-      if (point.dx.isFinite && point.dy.isFinite) {
-        canvas.drawCircle(point, 4, pointPaint);
-      }
-    }
-
-    // Highlight last point
-    if (points.isNotEmpty) {
-      final lastPoint = points.last;
-      if (lastPoint.dx.isFinite && lastPoint.dy.isFinite) {
-        final highlightPaint = Paint()
-          ..color = SyntrakColors.accent
-          ..style = PaintingStyle.fill;
-        canvas.drawCircle(lastPoint, 6, highlightPaint);
-
-        // Draw vertical line
-        final linePaint = Paint()
-          ..color = SyntrakColors.textPrimary.withOpacity(0.3)
-          ..strokeWidth = 1;
-        canvas.drawLine(
-          Offset(lastPoint.dx, 0),
-          Offset(lastPoint.dx, size.height),
-          linePaint,
-        );
-      }
-    }
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) => false;
 }
