+++ frontend/lib/screens/profile/widgets/profile_activities_search_bar.dart
@@ -1,0 +1,65 @@
+import 'package:flutter/material.dart';
+import 'package:syntrak/core/theme.dart';
+
+class ProfileActivitiesSearchBar extends StatelessWidget {
+  const ProfileActivitiesSearchBar({
+    super.key,
+    required this.controller,
+    required this.onClear,
+  });
+
+  final TextEditingController controller;
+  final VoidCallback onClear;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: SyntrakColors.surface,
+      padding: const EdgeInsets.fromLTRB(
+        SyntrakSpacing.md,
+        SyntrakSpacing.md,
+        SyntrakSpacing.md,
+        SyntrakSpacing.sm,
+      ),
+      child: Container(
+        height: 44,
+        decoration: BoxDecoration(
+          color: SyntrakColors.surfaceVariant,
+          borderRadius: BorderRadius.circular(12),
+        ),
+        child: TextField(
+          controller: controller,
+          style: SyntrakTypography.bodyMedium.copyWith(
+            color: SyntrakColors.textPrimary,
+          ),
+          decoration: InputDecoration(
+            hintText: 'Search activities...',
+            hintStyle: SyntrakTypography.bodyMedium.copyWith(
+              color: SyntrakColors.textTertiary,
+            ),
+            prefixIcon: Icon(
+              Icons.search,
+              color: SyntrakColors.textTertiary,
+              size: 22,
+            ),
+            suffixIcon: controller.text.isNotEmpty
+                ? IconButton(
+                    icon: Icon(
+                      Icons.close,
+                      color: SyntrakColors.textSecondary,
+                      size: 20,
+                    ),
+                    onPressed: onClear,
+                  )
+                : null,
+            border: InputBorder.none,
+            contentPadding: const EdgeInsets.symmetric(
+              horizontal: SyntrakSpacing.md,
+              vertical: 12,
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
